+++ CHANGES.md
@@ -99,6 +99,8 @@ To be released.
  -  (Libplanet.Stun) Added `TurnClient.Create()` static method.  [[#2219]]
  -  (Libplanet.Explorer) Added `CurrencyType` class.  In GraphQL, it corresponds
     to `Currency` type.  [[#2149], [#2227]]
+ -  (Libplanet.Explorer) Added `FungibleAssetValueType` class.  In GraphQL,
+    it corresponds to `FungibleAssetValue` type.  [[#2149], [#2227]]
  -  (Libplanet.Explorer) Added `PolicyType` class.  In GraphQL, it corresponds
     to `Policy` type.  [[#2149], [#2227]]